+++ core/urls.py
@@ -38,5 +38,6 @@ urlpatterns = [
     path('programs/', include('programs.urls', namespace='programs')),
     path('app_services/', include('app.service_urls', namespace='app_services')),
     path('sessions/', include('user_sessions.urls', namespace='sessions')),
-    path('topics/', include('topics.urls', namespace='topics'))
+    path('topics/', include('topics.urls', namespace='topics')),
+    path('', include('password_reset.urls', namespace='password_reset'))
 ]
