+++ pages/urls.py
@@ -3,14 +3,9 @@ from pages import views
 
 app_name = 'pages'
 
-
 urlpatterns = [
     path('', views.HomeView.as_view(), name='index'),
     path('login/', views.LoginView.as_view(), name='login'),
     path('logout/', views.LogoutView.as_view(), name='logout'),
     path('dashboard/', views.DashboardView.as_view(), name='dashboard'),
-    path('password-reset/', views.PasswordResetView.as_view(), name='password_reset'),
-    path('password-reset/done/', views.PasswordResetDoneView.as_view(), name='password_reset_done'),
-    path('password-reset/confirm/<str:token>/', views.PasswordResetConfirmView.as_view(), name='password_reset_confirm'),
-    path('password-reset/complete/', views.PasswordResetCompleteView.as_view(), name='password_reset_complete')
 ]
