+++ pages/views.py
@@ -46,30 +46,3 @@ class DashboardView(LoginRequiredMixin, RedirectView):
         if roles.count() == 1:
             return reverse_lazy('groups:group_dashboard', kwargs=dict(group_id=roles.first().group_id))
         return reverse_lazy('groups:my_groups')
-
-
-class PasswordResetView(TemplateView):
-    template_name = 'pages/password_reset.html'
-
-class PasswordResetDoneView(TemplateView):
-    template_name = 'pages/password_reset_done.html'
-
-    def post(self, request, *args, **kwargs):
-        # TODO: generar token para cambio de constraseña
-        # TODO: guardar en base de datos
-        # TODO: enviar email
-        return super(TemplateView, self).render_to_response(self.get_context_data())
-
-class PasswordResetConfirmView(TemplateView):
-    # TODO: comprobar token
-    # TODO: mostrar formulario de cambio constraseña
-    # TODO: validar nueva contraseña y actualizar en base de datos.
-    template_name = 'pages/password_reset_confirm.html'
-
-
-class PasswordResetCompleteView(TemplateView):
-    # TODO: mensaje de confirmacion de contraseña actualizada.
-    template_name = 'pages/password_reset_complete.html'
-
-    def post(self, request, *args, **kwargs):
-        return super(TemplateView, self).render_to_response(self.get_context_data())
