+++ password_reset/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class PasswordResetConfig(AppConfig):
+    name = 'password_reset'
