+++ templates/pages/login.html
@@ -62,7 +62,7 @@
                   </form>
 
                   <div class="mt-3 text-center">
-                    <a href="{% url 'pages:password_reset' %}">¿Olvidó su contraseña?</a>
+                    <a href="{% url 'password_reset:password_reset' %}">¿Olvidó su contraseña?</a>
                   </div>
                 </div>
               </div>
