+++ templates/password_reset/password_reset.html
@@ -1,8 +1,8 @@
-{% extends 'pages/layout.html' %}
+{% extends 'password_reset/layout.html' %}
 
 {% block content %}
     <h3 class="text-center my-4">¿Olvidó su contraseña?</h3>
-    <form action="{% url 'pages:password_reset_done' %}" method="post">
+    <form action="{% url 'password_reset:password_reset_done' %}" method="post">
         {% csrf_token %}
         <div class="form-group">
             <label for="email">Email</label>
