+++ templates/password_reset/password_reset_confirm.html
@@ -1,11 +1,11 @@
-{% extends 'pages/layout.html' %}
+{% extends 'password_reset/layout.html' %}
 
 {%block content%}
     <!-- token es valido -->
     <div class="row">
         <div class="col">
             <h3 class="text-center my-4">Cambiar su clave</h3>
-            <form action="{% url 'pages:password_reset_complete' %}" method="post">
+            <form action="{% url 'password_reset:password_reset_complete' %}" method="post">
                 {% csrf_token %}
                 <div class="form-group">
                     <label for="password">Clave</label>
