+++ templates/password_reset/password_reset_done.html
@@ -1,4 +1,4 @@
-{% extends 'pages/layout.html' %}
+{% extends 'password_reset/layout.html' %}
 
 
 {%block content%}
